+++ src/redux/actions/kb-credentials-users/post-kb-credentials-user-success.js
@@ -1,6 +1,5 @@
 export const POST_KB_CREDENTIALS_USER_SUCCESS = 'POST_KB_CREDENTIALS_USER_SUCCESS';
 
-export const postKBCredentialsUserSuccess = userData => ({
+export const postKBCredentialsUserSuccess = () => ({
   type: POST_KB_CREDENTIALS_USER_SUCCESS,
-  payload: userData,
 });