+++ src/redux/epics/post-kb-credentials-user.js
@@ -10,6 +10,7 @@ import {
   POST_KB_CREDENTIALS_USER,
   postKBCredentialsUserSuccess,
   postKBCredentialsUserFailure,
+  getKBCredentialsUsers,
 } from '../actions';
 
 export default ({ kbCredentialsUsersApi }) => (action$, state$) => {
@@ -22,6 +23,7 @@ export default ({ kbCredentialsUsersApi }) => (action$, state$) => {
         .assignUser(state$.value.okapi, credentialsId, userData)
         .pipe(
           map(postKBCredentialsUserSuccess),
+          map(() => getKBCredentialsUsers(credentialsId)),
           catchError(errors => of(postKBCredentialsUserFailure({ errors }))),
         );
     }),