+++ src/routes/settings-assigned-users-route/settings-assigned-users-route.js
@@ -24,22 +24,9 @@ const propTypes = {
   assignedUsers: KbCredentialsUsers.kbCredentialsUsersReduxStateShape.isRequired,
   deleteKBCredentialsUser: PropTypes.func.isRequired,
   getKBCredentialsUsers: PropTypes.func.isRequired,
-  getUserGroups: PropTypes.func.isRequired,
   kbCredentials: KbCredentials.KbCredentialsReduxStateShape,
   match: ReactRouterPropTypes.match.isRequired,
   postKBCredentialsUser: PropTypes.func.isRequired,
-  userGroups: PropTypes.shape({
-    errors: PropTypes.arrayOf(PropTypes.shape({
-      title: PropTypes.string.isRequired,
-    })).isRequired,
-    hasFailed: PropTypes.bool.isRequired,
-    hasLoaded: PropTypes.bool.isRequired,
-    isLoading: PropTypes.bool.isRequired,
-    items: PropTypes.arrayOf(PropTypes.shape({
-      group: PropTypes.string.isRequired,
-      id: PropTypes.string.isRequired,
-    })).isRequired,
-  }).isRequired,
 };
 
 const SettingsAssignedUsersRoute = ({
@@ -49,16 +36,10 @@ const SettingsAssignedUsersRoute = ({
   assignedUsers,
   kbCredentials,
   match: { params: { kbId } },
-  userGroups,
-  getUserGroups,
 }) => {
   useEffect(() => {
     getKBCredentialsUsers(kbId);
   }, [getKBCredentialsUsers, kbId]);
-
-  useEffect(() => {
-    getUserGroups();
-  }, [getUserGroups]);
 
   const [
     alreadyAssignedMessageDisplayed,
@@ -95,37 +76,14 @@ const SettingsAssignedUsersRoute = ({
       }));
   }, [assignedUsers.errors]);
 
-  const getPatronGroupNameById = id => {
-    return userGroups.items.find(userGroup => userGroup.id === id)?.group;
-  };
 
   const getFormattedUserData = user => {
-    const {
-      patronGroup,
-      username,
-      id,
-      personal: {
-        firstName,
-        middleName,
-        lastName,
-      },
-    } = user;
-
-    const attributes = {
-      credentialsId: kbId,
-      patronGroup: getPatronGroupNameById(patronGroup),
-      lastName,
-    };
-
-    if (username) attributes.userName = username;
-    if (firstName) attributes.firstName = firstName;
-    if (middleName) attributes.middleName = middleName;
+    const { id } = user;
 
     return {
       data: {
         id,
-        type: 'assignedUsers',
-        attributes,
+        credentialsId: kbId,
       },
     };
   };
@@ -139,11 +97,10 @@ const SettingsAssignedUsersRoute = ({
   };
 
   const assignedUsersLoaded = assignedUsers.hasLoaded || assignedUsers.hasFailed;
-  const userGroupsLoaded = userGroups.hasLoaded || userGroups.hasFailed;
 
   return (
     <>
-      {assignedUsersLoaded && userGroupsLoaded
+      {assignedUsersLoaded
         ? (
           <View
             requestIsPending={assignedUsers.isLoading}